Codex: Rename proxy protocol package to server.
Move internal/protocol to internal/server and update bootstrap imports to match the new package name. Sync the OpenClaw codebase map with the new module path.
